Update FLoops to the version that invokes init in loop (#52)
As of https://github.com/JuliaFolds/FLoops.jl/pull/73 the init clauses
are evaluated in the loop body whenever required. This is not exercised in
FoldsCUDA.jl yet (supposedly). But, since FoldsCUDA.jl is one of the
most type-instability-sensitive FLoops.jl users, it makes sense to
start using/testing this new unreleased version.